requirements.txt
python-dotenv>=1.0.0,<2.0.0
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
pydantic>=2.10.0,<3.0.0
mlflow>=2.18.0,<3.0.0
scikit-learn>=1.5.0,<2.0.0
numpy>=2.0.0,<3.0.0
pandas>=2.2.0,<3.0.0
shapely>=2.0.0,<3.0.0
xgboost>=2.1.0,<3.0.0
joblib>=1.4.0,<2.0.0
